Keep session alive when command exits
Previously, whenever the command exited, the session would be closed.
This should no longer be the case. A session being canceled/closed
should terminate a running command process, but a command process
exiting should not end a session.

This enables use of a single session with multiple different commands,
as long as the data channel is alive.